requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf_spectacular
drf_extensions
psycopg2-binary
django-simple-history
pillow